# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Security AI Agent",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/chat": {
      "post": {
        "tags": [
          "chat"
        ],
        "summary": "Api Chat",
        "description": "Single-shot chat: classify + dispatch + answer.",
        "operationId": "api_chat_api_v1_chat_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/src__api__routers__chat__ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/vulnscan/conversations": {
      "get": {
        "tags": [
          "scan-chat"
        ],
        "summary": "Api List Conversations",
        "operationId": "api_list_conversations_api_v1_vulnscan_conversations_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "scan-chat"
        ],
        "summary": "Api Create Conversation",
        "operationId": "api_create_conversation_api_v1_vulnscan_conversations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateConversationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/vulnscan/conversations/{conv_id}": {
      "get": {
        "tags": [
          "scan-chat"
        ],
        "summary": "Api Get Conversation",
        "operationId": "api_get_conversation_api_v1_vulnscan_conversations__conv_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "conv_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Conv Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "scan-chat"
        ],
        "summary": "Api Update Conversation",
        "operationId": "api_update_conversation_api_v1_vulnscan_conversations__conv_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "conv_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Conv Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateConversationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "scan-chat"
        ],
        "summary": "Api Delete Conversation",
        "operationId": "api_delete_conversation_api_v1_vulnscan_conversations__conv_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "conv_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Conv Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/vulnscan/conversations/{conv_id}/chat": {
      "post": {
        "tags": [
          "scan-chat"
        ],
        "summary": "Api Chat",
        "description": "Multi-turn dialog. Persists user + assistant, returns the updated\nconversation along with the (auto-extracted) ScanIntent if any.",
        "operationId": "api_chat_api_v1_vulnscan_conversations__conv_id__chat_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "conv_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Conv Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/src__api__routers__scan_chat__ChatRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChatResponse": {
        "properties": {
          "intent": {
            "type": "string",
            "enum": [
              "scan",
              "project",
              "web",
              "chat"
            ],
            "title": "Intent"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence"
          },
          "reply": {
            "type": "string",
            "title": "Reply"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/SourceRef"
            },
            "type": "array",
            "title": "Sources",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "intent",
          "confidence",
          "reply"
        ],
        "title": "ChatResponse"
      },
      "CreateConversationRequest": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title",
            "default": "新对话"
          },
          "model_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Id"
          }
        },
        "type": "object",
        "title": "CreateConversationRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SourceRef": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          },
          "snippet": {
            "type": "string",
            "title": "Snippet",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "title"
        ],
        "title": "SourceRef"
      },
...
      "UpdateConversationRequest": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "model_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Id"
          }
        },
        "type": "object",
        "title": "UpdateConversationRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "src__api__routers__chat__ChatRequest": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          },
          "history": {
            "items": {
              "type": "object"
            },
            "type": "array",
            "title": "History",
            "default": []
          },
          "model_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Id"
          },
          "conversation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conversation Id"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "ChatRequest"
      },
      "src__api__routers__scan_chat__ChatRequest": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          },
          "model_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Id"
          },
          "parse_intent": {
            "type": "boolean",
            "title": "Parse Intent",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "ChatRequest"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}